requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.9.0
APScheduler==3.11.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
Authlib==1.6.5
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.5
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
boto3==1.38.31
botocore==1.38.31
bottle==0.13.3
build==1.3.0
cachetools==6.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.3.0
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
cyclopts==4.2.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
discord==2.3.2
discord.py==2.5.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
durationpy==0.10
email-validator==2.3.0
emoji==2.15.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.115.12
fastmcp==2.13.0.2
feedparser==6.0.11
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
greenlet==3.2.3
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==1.0.3
langchain-anthropic==1.0.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.1
langchain-mcp-adapters==0.1.11
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langchainhub==0.1.21
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==2.1.1
langgraph-checkpoint-redis==0.1.2
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.32
lightweight-charts==2.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.20.0
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
more-itertools==10.8.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.1
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.23.2
openai==2.1.0
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
ply==3.11
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
proxy_tools==0.1.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.10.1
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pynamodb==6.1.0
pyparsing==3.2.3
pypdf==6.1.3
pyperclip==1.11.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-iso639==2025.2.18
python-kis==2.1.3
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-ulid==3.1.0
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.0.0
RapidFuzz==3.14.3
redis==6.4.0
redisvl==0.10.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-rst==1.3.2
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
seaborn==0.13.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit-aggrid==1.1.5.post1
streamlit-lightweight-charts==0.7.20
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.15
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.25.0
